backend/detector/requirements.txt
pillow
paho-mqtt
